feat(CMakeLists.txt): fix add_custom_command for bison/flex to be compatible with cmake-3.0
CMP0050
-------

Disallow add_custom_command SOURCE signatures.

CMake 2.8.12 and lower allowed a signature for ``add_custom_command()``
which specified an input to a command.  This was undocumented behavior.
Modern use of CMake associates custom commands with their output, rather
than their input.

The OLD behavior for this policy is to allow the use of
``add_custom_command()`` SOURCE signatures.  The NEW behavior for this
policy is to issue an error if such a signature is used.

This policy was introduced in CMake version 3.0.
CMake version 3.0.0 warns when the policy is not set and uses
OLD behavior.  Use the cmake_policy command to set it to OLD or
NEW explicitly.