fix(dsolvers): move initialize_constraints into assert from push
close #56

When we process a SMT2 file, we have the following assumptions about the
order of API instructions:

 - All `inform` instructions come first
 - The first `push` comes after that
 - We have a sequence of `(assert+;check)+` following it.

Exploiting these assumptions, we place a call to
'initialize_constraints' function inside of pushBacktrackPoint().

This assumption is not valid anymore when dReal is used via API. A
typical example is:

 - `inform`
 - `push; assert; check`
 - `inform`
 - `push; assert; check`
 - `pop;pop;delete_ctx`

The previous code fails when we process the second 'assert' since m_ctrs
does not have information about 2nd inform.

This fix moves a call to 'initialize_constraints' from
pushBacktrackPoint() to assert(). We introduce a Boolean flag
'm_need_init' which controls the initialization process.