fix(util/ibex_enode.cpp): memory leak
The following is a part of ibex_ExprCtr.h code:

    inline const ExprCtr& ExprNode::operator>=(const ExprNode& right) const
    {
      return *new ExprCtr(right.is_zero()? *this : *this-right, GEQ);
    }

When right is zero, right will not be added to the final expression, and
will not be freed in the end as a result. This is one source of memory
leak.

This commit is a workaround for the problem. We have a static variable
(zero : ExprConstant) which represent a constant 0.0. When the
right-hand-side of a constraint is zero, we use the variable zero,
instead of calling translate_enode_to_exprnode for the right-hand-side
of a constraint.