fix(opensmt/parsers/smt2/smt2parser.yy): do not process assert in parsing time
close #144

We used to handle simple assertions such as (assert (<= x 10))
specially. First we extract domain information from an assertion, then
we do not add the assertion to the egraph.

However, this shouldn't be done statically in parsing time because an
assertion can be pushed and popped dynamically in solving time.