feat(dsolvers/nra_solver.cpp): In assertLit, treat special forms specially
When we have a simple comparison between a variable and a constant such
as x <= 10 or y == 10, we handle it specially in nra_solver::assertLit
method.

Normally, we add a coresponding constraint into a `m_stack` member
variable and check the consistency of stack later in nra_solver::check
command. However, when we detect special forms such as mentioned above,
we try to prune a current box and determine whether we need to add the
corresponding constraint into a stack or not.