feat(constraint): forallt constraint holds a nonlinear constraint
We take a list of mode invariants from a hybrid system model. Each line
is corresponding with a forallt constraint. For now, we restrict a
forallt constraint to be a nonlinear constraint. It cannot be a
disjunction or conjunction.

Note that it's not a big problem not to support a conjunction since each
conjunct can be represented as a separate forallt constrinat. Therefore,
the only limitation is that we do not allow users to use disjunctive
mode invariant.