refactor(contractor/contractor_capd4.cpp): have a list of ctcs for invariants
Previously, we create a contractor for mode invariants when a CAPD4
contractor is created. This turns out to be problematic because there
are unasserted/asserted forallt literals and we need to get the
information when we do *check*.

In the new code, we prepare a list of contractors for each forallt
constrains in construction time. Then, when we check, we lookup the
polarity of each forallt literal and only used the asserted ones.