feat(contractor/contractor_capd4): relax check_invariant
Previously, we prune out the given interval iv or a representation s
using an invariant. But it turns out that it causes problems in CAPD4
since it maintains not the interval vector but either doubleton or
tripleton set. For example, a doubleton set consists of x + C*r0 + B * r
where

 - the vector x is a center,
 - the matrices C and B are coordinate systems,
 - the vector r0 is a product of intervals and represents the set in a given coordinate system,
 - the vector r is used mainly to store computational errors

In this commit, we remove the shinking part of invariant checking. When
the current input is violating the given invariant (i.e. the
intersection between the input and invariant is an empty set), then it
returns false. Otherwise, it simply returns true instead of pruning the
input.

close #197