feat(opensmt/smtsolvers/CoreSMTSolver): add undo_unchecked_assertions
Related Issue: https://github.com/dreal/dreal3/issues/210

I think there is a problem in the current way of handling push/pop in
the OpenSMT implementation. When an 'assert' is called, it saves the
assertion in the egraph.assertions and delays its processing until a
check is executed. If a push operation is executed before a check, the
previous assertion is not saved and, as a result, a pop operation will
remove the assertion which was asserted before a push is called.

For example, consider the following scenario.

    assert(F1)
    push()
    assert(F2)
    check()     -- (1)
    pop()
    check()     -- (2)

In this scenario, the first check (1) should check F1 /\ F2 and the
second one (2) should check F1. However, in the current implementation,
pop() removes both of F1 and F2 from the stack and let the second
check() checks nothing.

This patch adds undo_unchecked_assertions ( : std::vector<Enode *> ) and
save the current unchecked_assertions at push, and restore it when a
corresponding pop is called.