feat(util/ibex_enode): change the type of var_map
Previously, we used the type map<string, Variable const> for var_map
which is used as a parameter in translate_enode_to_exprnode and
translate_enode_to_exprctr, and as a return type of build_var_map
function.

By this commit, it is changed to be map<string, ExprSymbol
const *>. This is because ibex::Variable class is a mere wrapper of
ibex::ExprSymbol const * and we want to avoid synchronization in the
global varibalbes() : ibex::NodeMap.