feat(util/thread_local.h): add DREAL_THREAD_LOCAL
We are using C++11's thread_local keyword in the code base. However,
Apple's Clang does not support it (as of 2016 Aug, in XCode 7.3).

In this commit, we introduce `DREAL_THREAD_LOCAL` and replace all
previous occurrences of `thread_local`. In src/CMakeLists.txt, we check
whether a compiler/library supports the thread_local keyword or not, and
define COMP_HAS_THREAD_LOCAL macro in compatibility.h
file. `src/util/thread_local.h` defines DREAL_THREAD_LOCAL based on
COMP_HAS_THREAD_LOCAL. For now, the mapping is

    COMP_HAS_THREAD_LOCAL is defined     => thread_local
    COMP_HAS_THREAD_LOCAL is not defined =>               (nothing)

As a result, it's not *safe* to use a variable defined with
`DREAL_THREAD_LOCAL` in a multi-thread environment. We need to disable
multi-thread support if thread_local is not available.