feat(contractor/contractor_forall.cpp): drop f(y) in CE pruning
That is, do not use a constraint f(y) which only includes universally
quantified variables, not existentially quantified variables.

This is still problematic since when we have a false counterexample, CE,
such that f(CE) holds (note: to be a true counterexmaple, it should be
the case that !f(CE) holds), we should add a given input box B into the
returned box set.

Two possible fixes are:

1) Make sure that we always have true counterexamples.
2) Evaluate f(CE) and add a given input box if f(CE) = true.