feat: change behavior of unavailable features
We have features relying on optional third-party packages (CLP, GLPK,
NLOPT). Previously, we disable corresponding command-line options and
API functions when an optional package is not installed or configured
properly. For example, when `USE_GLPK=OFF` is used, `dReal --help`
does not show `--lp-icp` option and dReal API does not provide
set_lp(bool b) function.

This commit changes the behavior. Regardless of the installed/configured
packages, we provide the options for the features. However, we output
messages when a user uses those options. `dReal --help` also notifies
that certain options are not available. Here are examples:

From `dReal --help`:

     --polytope    use polytope contractor in IBEX (not available in
       this build, need CLP)

When setLP(bool) is called while GLPK is not installed:
     --lp-icp option is not available in this build. "
     To use it, please configure dReal with -DUSE_GLPK=ON cmake option.