fix(CMakeLists.txt): add missing link libraries
Modify CMake logic to actually link the dreal and dop libraries to their
dependencies. This fixes the build when -Wl,--no-undefined is used
(which should always be usable aside from some rare, esoteric cases),
and should also fix the issue mentioned in #51 where dlopen'ing the
library requires first dlopen'ing its dependencies. It will also allow
drastically simplifying the link line of dreal.pc when built shared, so
that we aren't leaking our list of dependencies. (When built static,
there is of course no such thing as private linking; we'll still need
the full list in that case.)

Closes #316