fix(CMakeLists.txt): refactor how dylib is created
Refactor the logic to create the macOS .dylib to actually use a CMake
library rather than a custom command. This is a little bit shorter, and
not going behind CMake's back as far as how to create a library is
probably a good idea. Also, add the link flags that tell the linker
where to actually find the external libraries, which should fix creation
of the .dylib failing. (We use /dev/null as a dummy source file, since
the library must include at least one source. Portability isn't an issue
here since this logic is only happening on macOS, which had better have
/dev/null.)

Note that this removes a bunch of manually specified dependencies; these
aren't needed, as they are picked up already transitively via the dreal
target.