fix(dsolvers/nra_solver.cpp): always add constraint to the stack.
We used to skip adding a constraint to the stack if the constraint is
simple relational constraint in the form of (var rel-op constraint)
where rel-op = {<, <=, >, >=, =}. The intuition is that we can
transfer this constraint into the box by restricting it and we do not
need to keep the constraint anymore. This is an optimization.

However, this causes a problem when we have more than one
check-sat. Please take a look at the following example:

1 (set-logic QF_NRA)
2 (declare-fun x () Real)
3 (assert (> x 0.0))
4 (assert (< x 2.0))
5 (check-sat)
6 (assert (> x 5.0))
7 (check-sat)
8 (exit)

Between the first "check-sat" in line 5 and "assert" in line 6, dReal
first calls "inform x > 5.0" which resets the current box (m_cs.m_box)
since an inform command can introduce a new variable which changes the
size of the interval box.

Without this patch, we lose the information "0 < x < 2" between line 5
and 6. As a result, the second "check-sat" will get "delta-sat".

This patch essentially disables an optimization and always keep a
constraint in the stack. The box will still be reset at line 6, but we
have "0 < x < 2" in the stack. This will help dReal detect
inconsistency when it handles "x > 5".